+++ src/GafferVDBModule/GafferVDBModule.cpp
@@ -40,6 +40,7 @@
 #include "GafferVDB/LevelSetToMesh.h"
 #include "GafferVDB/MeshToLevelSet.h"
 #include "GafferVDB/PointsGridToPoints.h"
+#include "GafferVDB/PointsToLevelSet.h"
 #include "GafferVDB/SphereLevelSet.h"
 #include "GafferBindings/DependencyNodeBinding.h"
 
@@ -54,5 +55,6 @@ BOOST_PYTHON_MODULE( _GafferVDB )
 	GafferBindings::DependencyNodeClass<LevelSetOffset>();
 	GafferBindings::DependencyNodeClass<PointsGridToPoints>();
 	GafferBindings::DependencyNodeClass<SphereLevelSet>();
+	GafferBindings::DependencyNodeClass<PointsToLevelSet>();
 
 }